requirements.txt
google-genai>=1.0.0
sendgrid>=6.11.0
jinja2>=3.1.0
pyyaml>=6.0.0
